Fix bug in InfiniteRepeaterRetryStrategy.
If the strategy is used to get the wait time of retry N right away,
instead of starting at 1, the lastWaitTime would always be zero
(00:00:00) instead of the actual wait time suggested by the inner
strategy as the last wait time. In that case, discover the last wait
time by asking the inner strategy for it.